refactor components, add images and videos in assets, add a uniform navbar to every pages
