+++ client/src/assets/images/HopBox_Logo.png
