+++ client/src/components/LandingComponents/CoverSection/CoverElements.jsx
