+++ client/src/components/LandingComponents/InfoSection/index.jsx
