+++ client/src/components/LandingComponents/Services/index.jsx
@@ -1,12 +1,11 @@
 import React, { useState } from 'react';
-import word from '../../assets/hopbox word.png';
-import demo from '../../assets/coverobject.jpeg';
-import object1 from '../../assets/object1.png';
+import word from '../../../assets/images/hopbox word.png';
+import demo from '../../../assets/images/coverobject.jpeg';
+import object1 from '../../../assets/images/object1.png';
 import TextField from '@material-ui/core/TextField';
-import { Button } from '../ButtonElements';
 import { ServicesContainer, ServicesH1, ServicesWrapper, ServicesH2, ServicesP, Imgbg, Divider } from './ServicesElements';
 import { Imgthingie, Imgword, ServicesBox, SubmitButton, ServicesH3, SignUpWrapper } from './ServicesElements';
-import * as smoketest from '../../api';
+import * as smoketest from '../../../api';
 
 
 const Services = () => {
